move preloading communities to its own hook/util
