+++ src/components/layout/AppSidebar.tsx
@@ -17,10 +17,7 @@ import { useIsMobile } from "@/hooks/useIsMobile";
 import { UserStatusIndicator } from "@/components/user/UserStatusIndicator";
 import { UserMenu } from "@/components/user/UserMenu";
 import { logger } from "@/lib/logger";
-import { useQueryClient } from "@tanstack/react-query";
-import { useNostr } from "@nostrify/react";
-import { useEventCache } from "@/hooks/useEventCache";
-import { fetchMessages } from "@/hooks/useMessages";
+import { usePreloadCommunity } from "@/hooks/usePreloadCommunity";
 import { useState, useEffect, useCallback, useMemo, useRef } from "react";
 import {
   DndContext,
@@ -290,9 +287,7 @@ export function AppSidebar({
   const { user } = useCurrentUser();
   const author = useAuthor(user?.pubkey || '');
   const metadata = author.data?.metadata;
-  const queryClient = useQueryClient();
-  const { nostr } = useNostr();
-  const { cacheEvents } = useEventCache();
+  const { preloadCommunity } = usePreloadCommunity();
 
   // State for sophisticated animation
   const [launchingCommunity, setLaunchingCommunity] = useState<string | null>(null);
@@ -393,77 +388,12 @@ export function AppSidebar({
     };
   }, []);
 
-  // Community preloading function - restored from commit baecfd3
-  const preloadCommunity = useCallback(async (communityId: string) => {
-    // Skip if already selected
-    if (communityId === selectedCommunity) {
-      logger.log('[DEBUG PRELOAD] Skipping preload - community already selected:', communityId);
-      return;
-    }
 
-    logger.log('[DEBUG PRELOAD] Starting community preload:', communityId);
-
-    try {
-      // Check if data is already cached and fresh (5 minutes)
-      const CACHE_DURATION = 5 * 60 * 1000;
-      const channelsQueryKey = ['channels', communityId];
-      // Use the exact same query key structure as useMessages: ['messages', communityId, channelId, approvedMembers?.size]
-      // Most communities don't require approval, so approvedMembers will be null (no filtering)
-      // This is the most common case and what we should preload for
-      const messagesQueryKey = ['messages', communityId, 'general', null];
-      
-      const channelsState = queryClient.getQueryState(channelsQueryKey);
-      const messagesState = queryClient.getQueryState(messagesQueryKey);
-      
-      const now = Date.now();
-      const isChannelsFresh = channelsState?.dataUpdatedAt && (now - channelsState.dataUpdatedAt) < CACHE_DURATION;
-      const isMessagesFresh = messagesState?.dataUpdatedAt && (now - messagesState.dataUpdatedAt) < CACHE_DURATION;
-
-      logger.log('[DEBUG PRELOAD] Cache status:', {
-        communityId,
-        channels: { fresh: isChannelsFresh, lastUpdate: channelsState?.dataUpdatedAt ? new Date(channelsState.dataUpdatedAt).toISOString() : 'never' },
-        messages: { fresh: isMessagesFresh, lastUpdate: messagesState?.dataUpdatedAt ? new Date(messagesState.dataUpdatedAt).toISOString() : 'never' }
-      });
-
-      // Preload channels if not fresh - uses EXACT same query as useChannels
-      if (!isChannelsFresh) {
-        logger.log('[DEBUG PRELOAD] Prefetching channels for:', communityId);
-        queryClient.prefetchQuery({
-          queryKey: channelsQueryKey,
-          staleTime: CACHE_DURATION,
-        });
-      } else {
-        logger.log('[DEBUG PRELOAD] Channels already fresh for:', communityId);
-      }
-
-      // Preload general channel messages if not fresh - uses EXACT same query as useMessages
-      if (!isMessagesFresh) {
-        logger.log('[DEBUG PRELOAD] Prefetching general messages for:', communityId);
-        queryClient.prefetchQuery({
-          queryKey: messagesQueryKey,
-          queryFn: async ({ signal }) => {
-            logger.log('[DEBUG PRELOAD] Executing fetchMessages query for general channel:', communityId);
-            const result = await fetchMessages(communityId, 'general', nostr, cacheEvents, signal, null);
-            logger.log('[DEBUG PRELOAD] Fetched', result.events.length, 'messages for general channel:', communityId);
-            return result.events;
-          },
-          staleTime: CACHE_DURATION,
-        });
-      } else {
-        logger.log('[DEBUG PRELOAD] General messages already fresh for:', communityId);
-      }
-
-      logger.log('[DEBUG PRELOAD] Community preload completed:', communityId);
-    } catch (error) {
-      // Silently handle preload errors
-      logger.log('[DEBUG PRELOAD] Community preload failed:', communityId, error);
-    }
-  }, [selectedCommunity, queryClient, nostr, cacheEvents]);
 
   // Add mouse down handler for community interactions
   const handleCommunityMouseDown = useCallback((communityId: string) => {
-    preloadCommunity(communityId);
-  }, [preloadCommunity]);
+    preloadCommunity(communityId, selectedCommunity || undefined);
+  }, [preloadCommunity, selectedCommunity]);
 
   // Memoize the community list rendering to avoid unnecessary re-renders
   const communityListContent = useMemo(() => {
